+++ tensorlayer/logging/__init__.py
@@ -9,14 +9,14 @@ layers that allow user to apply ``tf.nn.lrn`` on ``network.outputs``.
 More functions can be found in `TensorFlow API <https://www.tensorflow.org/versions/master/api_docs/index.html>`__.
 """
 
+from tensorlayer.lazy_imports import LazyImport
+
 from .tl_logging import *
 
-from .contrib import *
+# Lazy Imports
+contrib = LazyImport("tensorlayer.logging.contrib")
 
 __all__ = [
-    # additional modules
-    'contrib',
-
     # tl_logging
     'DEBUG',
     'debug',